Moved stub form components in dedicated folders
