+++ gui/src/components/stubForm/formHelper/ExampleSelector.vue
