+++ gui/src/components/stubForm/formHelper/HttpMethodSelector.vue
