+++ gui/src/components/stubForm/formHelper/HttpStatusCodeSelector.vue
