+++ gui/src/components/stubForm/formHelper/LineEndingSelector.vue
