+++ gui/src/components/stubForm/formHelper/ResponseBodyHelper.vue
