+++ gui/src/components/stubForm/formHelper/ScenarioSelector.vue
