+++ gui/src/components/stubForm/formHelper/SetDynamicMode.vue
