+++ gui/src/components/stubForm/formHelper/SetQuery.vue
