+++ gui/src/views/StubForm.vue
@@ -61,8 +61,8 @@ import { simpleEditorThreshold } from "@/constants/technical";
 import { handleHttpError } from "@/utils/error";
 import yaml from "js-yaml";
 import { clearIntermediateStub, getIntermediateStub } from "@/utils/session";
-import FormHelperSelector from "@/components/stub/FormHelperSelector.vue";
-import StubFormButtons from "@/components/stub/StubFormButtons.vue";
+import FormHelperSelector from "@/components/stubForm/formHelper/FormHelperSelector.vue";
+import StubFormButtons from "@/components/stubForm/StubFormButtons.vue";
 import SimpleEditor from "@/components/simpleEditor/SimpleEditor.vue";
 import { error } from "@/utils/toast";
 import { useStubsStore } from "@/store/stubs";
